Make Kubernetes module, split code out
