+++ src/kubernetes/mod.rs
@@ -1,0 +1,3 @@
+pub mod controller;
+pub mod operations;
+pub mod reconciler;
