+++ src/prometheus/alert.rs
@@ -1,10 +1,12 @@
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, HashMap};
 
 use color_eyre::{eyre::eyre, Result};
 use serde::{Deserialize, Serialize};
 
+use crate::service_alerts::{Alerts, ServiceAlertSpec};
+
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub struct Alerts {
+pub struct PromAlerts {
     pub groups: Vec<AlertGroup>,
 }
 
@@ -45,10 +47,10 @@ pub struct Annotations {
     pub description: String,
 }
 
-impl TryFrom<Alerts> for BTreeMap<String, String> {
+impl TryFrom<PromAlerts> for BTreeMap<String, String> {
     type Error = color_eyre::Report;
 
-    fn try_from(value: Alerts) -> Result<Self, Self::Error> {
+    fn try_from(value: PromAlerts) -> Result<Self, Self::Error> {
         // owner should be a unique identifier, so at least for now we can use
         // it as the key for our `BTreeMap`
         let identifier = match value.groups.first() {
@@ -62,6 +64,63 @@ impl TryFrom<Alerts> for BTreeMap<String, String> {
         let yaml_string = serde_yaml::to_string(&value)?;
 
         Ok(BTreeMap::from([(identifier, yaml_string)]))
+    }
+}
+
+/// FIXME: This should be replaced with a generated/converted when possible.
+/// Once this is marked as DEAD_CODE then we are good to go!
+const PLACEHOLDER_VALUE: &str = "PLACEHOLDER";
+
+impl TryFrom<ServiceAlertSpec> for PromAlerts {
+    type Error = color_eyre::Report;
+
+    fn try_from(value: ServiceAlertSpec) -> Result<Self, Self::Error> {
+        use crate::prometheus::alert::*;
+
+        let mut alerts = PromAlerts {
+            groups: Vec::with_capacity(value.alerts.len()),
+        };
+
+        if let Some(replica_alerts) = value.alerts.get(&Alerts::ReplicaCount) {
+            let replica_rules = replica_alerts
+                .iter()
+                .map(|conf| AlertRules {
+                    alert: PLACEHOLDER_VALUE.into(),
+                    expr: PLACEHOLDER_VALUE.into(),
+                    for_: conf.for_.clone(),
+                    labels: Labels {
+                        severity: PrometheusSeverity::from(&conf.alert_with_labels),
+                        source: value.common_labels.get("origin").unwrap().into(),
+                        owner: value.common_labels.get("owner").unwrap().into(),
+                    },
+                    annotations: Annotations {
+                        summary: PLACEHOLDER_VALUE.into(),
+                        description: PLACEHOLDER_VALUE.into(),
+                    },
+                })
+                .collect();
+
+            alerts.groups.push(AlertGroup {
+                name: PLACEHOLDER_VALUE.into(),
+                rules: replica_rules,
+            })
+        }
+
+        Ok(alerts)
+    }
+}
+
+/// CRD Severities are currently part of a HashMap, so we need to grab them from
+/// that structure. Since we don't need to modify or consume the HashMap, I
+/// borrow it here.
+impl From<&HashMap<String, String>> for PrometheusSeverity {
+    fn from(value: &HashMap<String, String>) -> Self {
+        match value.get("severity").unwrap().as_str() {
+            "warning" => PrometheusSeverity::Warning,
+            "critical" => PrometheusSeverity::Critical,
+            "page" => PrometheusSeverity::Page,
+            _ => PrometheusSeverity::Warning,
+        }
     }
 }
 
@@ -89,7 +148,7 @@ groups:
 
     #[test]
     fn test_serialisation_happy_path() -> Result<()> {
-        let rust_repr = Alerts {
+        let rust_repr = PromAlerts {
             groups: vec![AlertGroup {
                 name: "example".into(),
                 rules: vec![AlertRules {
@@ -109,7 +168,7 @@ groups:
             }],
         };
 
-        let yaml_repr: Alerts = serde_yaml::from_str(SERIALIZED_PROM_ALERT)?;
+        let yaml_repr: PromAlerts = serde_yaml::from_str(SERIALIZED_PROM_ALERT)?;
         assert_eq!(yaml_repr, rust_repr);
 
         Ok(())
